requirement.txt
Package                       Version
----------------------------- -------
asgiref                       3.7.2
Django                        4.2.2
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
pip                           23.1.2
PyJWT                         2.7.0
pytz                          2023.3
setuptools                    65.5.0
sqlparse                      0.4.4
tzdata                        2023.3
